telegram_youtube_bot_DEPLOY_FACILE/telegram_youtube_bot/requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3

